Add files via upload
The main idea of the project is to develop a data mining application which is based on citizen complaint analysis. The data used for this application is a citizen complaint data of Detroit for the year 2016. It is provided by the Detroit Police Department on the open data portal for the city of Detroit. The results of this application would be helpful in detecting areas with more complaints from the citizen.